+++ ridi_oauth2/introspector/constants.py
@@ -1,0 +1,11 @@
+JWK_EXPIRES_MIN = 30
+
+
+class JWKKeyType:
+    RSA = 'RSA'
+    EC = 'EC'
+    OCT = 'oct'
+
+
+class JWKUse:
+    SIG = 'sig'