+++ ridi_oauth2/introspector/helpers.py
@@ -1,33 +1,31 @@
-from typing import Dict
-
 import jwt
+from jwt import InvalidTokenError
+from jwt.exceptions import InvalidKeyError
 
-from ridi_oauth2.introspector.dtos import AccessTokenInfo, JwtInfo
+from ridi_oauth2.introspector.dtos import AccessTokenInfo
 from ridi_oauth2.introspector.exceptions import InvalidJwtSignatureException, InvalidToken
-from ridi_oauth2.introspector.jwt_introspector import JwtIntrospector
+from ridi_oauth2.introspector.key_handler import KeyHandler
 
 
 class JwtIntrospectHelper:
     @staticmethod
-    def introspect(jwt_infos: Dict[str, JwtInfo], access_token: str) -> AccessTokenInfo:
+    def introspect(access_token: str) -> AccessTokenInfo:
         try:
             unverified_header = jwt.get_unverified_header(access_token)
+            unverified_payload = jwt.decode(access_token, verify=False)
         except jwt.InvalidTokenError:
             raise InvalidToken
 
-        kid = unverified_header.get('kid')
-        if not kid:
+        kid = unverified_header.get('kid', None)
+        client_id = unverified_payload.get('client_id', None)
+
+        if not kid or not client_id:
             raise InvalidJwtSignatureException
 
-        jwt_info = jwt_infos.get(kid)
-        if not jwt_info:
-            raise InvalidJwtSignatureException
-
-        introspector = JwtIntrospector(jwt_info=jwt_info, access_token=access_token)
-        result = introspector.introspect()
+        public_key = KeyHandler.get_public_key_by_kid(client_id, kid)
 
         try:
-            return AccessTokenInfo.from_dict(result)
-
-        except KeyError:
+            payload = jwt.decode(jwt=access_token, key=public_key, algorithms=unverified_header.get('alg'))
+            return AccessTokenInfo.from_dict(payload)
+        except (InvalidTokenError, InvalidKeyError):
             raise InvalidJwtSignatureException