+++ CHANGES.md
@@ -1,5 +1,12 @@
 Changelog
 =========
+1.0.0 (Oct 10st 2019)
+------------------
+- Change main logic to get public key from OAuth2 server
+    - to get public key, it sends request to OAuth2 server, and memorize that key until key expires.
+    - when getting public key, it use [JWKS](https://tools.ietf.org/html/rfc7517) type
+    - it makes able to adapt OAuth2 server's key changing dynamically and using multi key.  
+
 0.0.15 (Aug 1st 2019)
 ------------------
 - Add cryptography in package